Relocated parts of test_eqdsk.py to CODES (Integrated_Development internal repo)
Renamed testing folder
